GUI work and and operation stepping
Vectorized some code for splitting the steps up, added text above
boxes, added dynamic resizing of the entry and output matrixes, added a
window for future explanations of steps. Added buttons for the option
of stepping through the matrix operation piece by piece. Still need to
add a way to change stepping method based on the method of operation
choosen.